pyproject.toml
[project]
name = "open-arena"
version = "0.1.0"
description = "Autoresearch loop for reward R&D and model selection — between post-training evals and the next round of fine-tuning / RL."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{ name = "Athena-Reply" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Environment :: Console",
    "Operating System :: OS Independent",
]
dependencies = [
    "arize-phoenix-client>=2.6.0",
    "braintrust>=0.18.0",
    "click>=8.1.8",
    "datasets>=4.8.5",
    # Direct imports in src/datasets/dataset.py + evaluate.py — declared
    # explicitly so installs don't rely on these being pulled in
    # transitively (today they ship via synalinks / datasets / langchain,
    # but a future minor version bump elsewhere could drop them).
    "jinja2>=3.1",
    "keras-tuner>=1.4.8",
    "langfuse>=4.5.1",
    "langsmith>=0.7.38",
    "numpy>=2.0",
    "opik>=2.0.20",
    "pyarrow>=23.0.1",
    "pyyaml>=6.0",
    "synalinks>=0.8.11",
]

[project.optional-dependencies]
# `deep_eval` reward — wraps `deepeval.metrics` so any DeepEval metric
# (G-Eval, AnswerRelevancy, Hallucination, …) can be selected from YAML.
# Install with `uv sync --extra deepeval`.
deepeval = ["deepeval>=2.0.0"]

[project.scripts]
arena = "evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
# Top-level `evaluate.py` (the sweep entry point used by the `arena`
# console script) and `program.py` (the editable `build_program()`) are
# not under `src/`, so force-include them in the wheel.
force-include = { "evaluate.py" = "evaluate.py", "program.py" = "program.py" }
